Add Final TextRecord Interface
Parameter text is really the only class that is not final. Shell script
has the feature that a redirection can be found in place of a parameter,
signifying the end of all parameters. It would probably be best to extend
Redirection out of parameter, and rather than cancel the parsing, perform
an upgrade.